add local file system switch
