+++ eq-author/cypress/integration/authenticated/dashboard_spec.js
@@ -87,17 +87,16 @@ describe("dashboard", () => {
       cy.login();
     });
 
-      it("should delete the questionnaire", () => {
-        cy.get(testId("btn-delete-questionnaire")).click();
-        cy.get(testId("btn-delete-modal")).click();
-        cy.contains("You have no questionnaires").should("be.visible");
-      });
+    it("should delete the questionnaire", () => {
+      cy.get(testId("btn-delete-questionnaire")).click();
+      cy.get(testId("btn-delete-modal")).click();
+      cy.contains("You have no questionnaires").should("be.visible");
+    });
 
-      it("should display toast when questionnaire is deleted", () => {
-        cy.get(testId("btn-delete-questionnaire")).click();
-        cy.get(testId("btn-delete-modal")).click();
-        cy.contains("Questionnaire deleted").should("be.visible");
-      });
+    it("should display toast when questionnaire is deleted", () => {
+      cy.get(testId("btn-delete-questionnaire")).click();
+      cy.get(testId("btn-delete-modal")).click();
+      cy.contains("Questionnaire deleted").should("be.visible");
     });
   });
 });
